fix: ColorSelectorの修正 Close #12
