+++ src/constants/index.js
@@ -1,0 +1,13 @@
+const colors = [
+  { name: 'black', value: '#000000' },
+  { name: 'tomato', value: '#ff6347' },
+  { name: 'forestgreen', value: '#228b22' },
+  { name: 'cornflowerblue', value: '#6495ed' },
+  { name: 'navy', value: '#000080' },
+  { name: 'gold', value: '#ffd700' },
+  { name: 'orange', value: '#ffa500' },
+  { name: 'purple', value: '#9400d3' },
+  { name: 'pink', value: '#ffc0cb' }
+];
+
+export { colors };
